Allow translation overrides and remove references to SURFconext
See the upgrade instructions for a detailed description of the
changes. In summary, the following words are dynamically injected into
the translations in order to have a sane set of default translations
not specifically targeted to the educational world:

 - %suiteName%:
   - OpenConext (default)
   - SURFconext, ACMEconext (example overrides)

 - %organisationNoun%:
   - organisation/organisatie (default)
   - institution/instelling (example overrides)

 - %organisationNounPlural%:
   - organisations/organisaties (default)
   - institutions/instellingen (example overrides)

 - %accountNoun%:
   - organisation account/organisatieaccount (default)
   - institutional account/instellingsaccount (example overrides)

The WAYF help page is a special case because the questions and answers
all have their own translation key, and those questions can be
considered specific to SURFconext. So the help page is now simplified
to just render one single "translatable" blob of HTML. The default
content for the help page is "No help content available" - this
content can of course be overridden in overrides.[lang].php.

See: https://www.pivotaltracker.com/story/show/155467582